Adding 'Victory Points' and 'Total Points' ranking schemes.
rankings/schemes.go has been added to hopefully make it easier to add
new ways of ranking leagues in the future. The /about page has been
updated with a description of each of the now three supported ranking
systems.

The 'Power Points' ranking scheme that was used interchangeably with
'All-Play' has been removed to avoid confusion. In practice, they were
the same ranking scheme but since 'All-Play' has better name
recognition and a more obvious way of handling ties, it is used instead.